Extract KAPI HPA reconciliation into dedicated file
