+++ pkg/operation/botanist/component/kubeapiserver/kube_apiserver.go
@@ -101,13 +101,10 @@ type kubeAPIServer struct {
 
 func (k *kubeAPIServer) Deploy(ctx context.Context) error {
 	var (
-		needsHPA  = !k.values.Autoscaling.HVPAEnabled && k.values.Autoscaling.Replicas != nil && *k.values.Autoscaling.Replicas > 0
 		needsVPA  = !k.values.Autoscaling.HVPAEnabled
 		needsHVPA = k.values.Autoscaling.HVPAEnabled && k.values.Autoscaling.Replicas != nil && *k.values.Autoscaling.Replicas > 0
 
-		hpaTargetAverageUtilizationCPU    int32 = 80
-		hpaTargetAverageUtilizationMemory int32 = 80
-		vpaUpdateMode                           = autoscalingv1beta2.UpdateModeOff
+		vpaUpdateMode = autoscalingv1beta2.UpdateModeOff
 
 		horizontalPodAutoscaler = k.emptyHorizontalPodAutoscaler()
 		verticalPodAutoscaler   = k.emptyVerticalPodAutoscaler()
@@ -118,43 +115,8 @@ func (k *kubeAPIServer) Deploy(ctx context.Context) error {
 	if err := k.reconcilePodDisruptionBudget(ctx, podDisruptionBudget); err != nil {
 		return err
 	}
-
-	if needsHPA {
-		if _, err := controllerutils.GetAndCreateOrMergePatch(ctx, k.client, horizontalPodAutoscaler, func() error {
-			horizontalPodAutoscaler.Spec = autoscalingv2beta1.HorizontalPodAutoscalerSpec{
-				MinReplicas: &k.values.Autoscaling.MinReplicas,
-				MaxReplicas: k.values.Autoscaling.MaxReplicas,
-				ScaleTargetRef: autoscalingv2beta1.CrossVersionObjectReference{
-					APIVersion: appsv1.SchemeGroupVersion.String(),
-					Kind:       "Deployment",
-					// TODO: Replace with `deployment.Name` once the Deployment is generated by this component.
-					Name: v1beta1constants.DeploymentNameKubeAPIServer,
-				},
-				Metrics: []autoscalingv2beta1.MetricSpec{
-					{
-						Type: autoscalingv2beta1.ResourceMetricSourceType,
-						Resource: &autoscalingv2beta1.ResourceMetricSource{
-							Name:                     corev1.ResourceCPU,
-							TargetAverageUtilization: &hpaTargetAverageUtilizationCPU,
-						},
-					},
-					{
-						Type: autoscalingv2beta1.ResourceMetricSourceType,
-						Resource: &autoscalingv2beta1.ResourceMetricSource{
-							Name:                     corev1.ResourceMemory,
-							TargetAverageUtilization: &hpaTargetAverageUtilizationMemory,
-						},
-					},
-				},
-			}
-			return nil
-		}); err != nil {
-			return err
-		}
-	} else {
-		if err := kutil.DeleteObject(ctx, k.client, horizontalPodAutoscaler); err != nil {
-			return err
-		}
+	if err := k.reconcileHorizontalPodAutoscaler(ctx, horizontalPodAutoscaler); err != nil {
+		return err
 	}
 
 	if needsVPA {
@@ -192,7 +154,7 @@ func (k *kubeAPIServer) Deploy(ctx context.Context) error {
 					Type: autoscalingv2beta1.ResourceMetricSourceType,
 					Resource: &autoscalingv2beta1.ResourceMetricSource{
 						Name:                     corev1.ResourceCPU,
-						TargetAverageUtilization: &hpaTargetAverageUtilizationCPU,
+						TargetAverageUtilization: pointer.Int32(hpaTargetAverageUtilizationCPU),
 					},
 				},
 			}
@@ -227,7 +189,7 @@ func (k *kubeAPIServer) Deploy(ctx context.Context) error {
 				Type: autoscalingv2beta1.ResourceMetricSourceType,
 				Resource: &autoscalingv2beta1.ResourceMetricSource{
 					Name:                     corev1.ResourceMemory,
-					TargetAverageUtilization: &hpaTargetAverageUtilizationMemory,
+					TargetAverageUtilization: pointer.Int32(hpaTargetAverageUtilizationMemory),
 				},
 			})
 		}
@@ -370,10 +332,6 @@ func (k *kubeAPIServer) GetValues() Values {
 }
 func (k *kubeAPIServer) SetAutoscalingReplicas(replicas *int32) {
 	k.values.Autoscaling.Replicas = replicas
-}
-
-func (k *kubeAPIServer) emptyHorizontalPodAutoscaler() *autoscalingv2beta1.HorizontalPodAutoscaler {
-	return &autoscalingv2beta1.HorizontalPodAutoscaler{ObjectMeta: metav1.ObjectMeta{Name: v1beta1constants.DeploymentNameKubeAPIServer, Namespace: k.namespace}}
 }
 
 func (k *kubeAPIServer) emptyVerticalPodAutoscaler() *autoscalingv1beta2.VerticalPodAutoscaler {
